Developer-only support for HIP-over-CUDA
This commit adds an *undocumented* mechanism to test HIP Memory Kinds
support via AMD's HIP-over-CUDA (aka HIP_PLATFORM=nvcc).

One of the advertised features of HIP is that it includes
functionality expressing a *subset* of CUDA, and thus can be transformed to
CUDA via preprocess.  This allows one to incrementally convert CUDA to
HIP, having both in one's application simultaneously.  This lets one
complete the conversion on a CUDA GPU before moving to an AMD one.
So, the mechanism introduced in this commit can hypothetically be used
by client authors in the midst of such a transition.  However, I am
withholding documentation until/unless we commit to test this feature
to prevent "bitrot".