Fix API violations in testam.h
This commit eliminates invalid calls to `gex_Segment_QueryBound()`
from AM handler context, via `TEST_SEG()`, in the logic common to
`testgasnet` and `testlegacy`, as well as their C++ analogues.

In the case of calls for one's own segment, the resolution is to have
the test codes (`testgasnet.c` and `testlegacy.c`) store the value in
a global variable prior to running the associated AM tests.  For the
case of query on a remote segment for construction of a ReplyLong, the
solution deployed is to add the sender's segment address to the
payload of all Medium and Long AMs.

Other than the addition of the segment address to the payload
structure, no changes have been made to the portions of
`other/amx/testam.h` used by `other/amx/testam.c`.