Refactor gex_Segment_Attach()
This commit replaces the replicated per-conduit implementations of
`gex_Segment_Attach()`, under the name `gasnetc_Segment_Attach()`,
with a single implementation.  Where necessary, native conduits can
define `GASNETC_SEGMENT_ATTACH_HOOK` in their `gasnet_core_fwd.h`
to enable use of a hook for memory registration and exchange of
the resulting registration keys.

The template-conduit is updated with a sketch of the new hook, in
addition to the four concrete examples introduced in the native
conduits.

It is hoped that future work will re-implement `gex_Segment_Attach()`
in terms of `gex_Segment_Create()` and `gex_EP_PublishBoundSegment()`.
That may eliminate the attach-specific hook in favor of the
conduit-specific hooks in those two functions (once they are properly
refactored).

In addition to refactoring, this commit adds `GASNETI_TRACE_PRINTF()`
for calls to `gex_Segment_Attach()`, with type='O'.  This is the only
intentional behavioral change in this commit.