Rename/revise per-thread medium buffer allocation
This commit recharacterizes the pair of "loopback" medium buffers as
"per-thread" medium buffers, with renames of corresponding functions
and TLS struct fields.

Additionally, the allocation and free functions are enhanced with
debug checks to detect incorrect usage.

This work is in preparation for use of these buffers in the PSHM and
reference implementations of NPAM as an alternative to malloc/free.