Refactor gasnet_attach()
This commit clones the `gasnetc_attach()` implementation from
`template-conduit/gasnet_core.c` to become `gasneti_attach()` in
`gasnet_legacy.c` with a corresponding change to the prototype and
single caller in `gasnet.h`.  Other than the name change and cosmetic
changes to the trace output, the new function is textually identical
to the template.

The `gasnetc_attach()` implementations in all conduits (and the
template) are removed in favor of the new common implementation.
Additionally, the `static` qualifier is removed from every conduit's
`gasnetc_attach_primary()` to permit calls from `gasneti_attach()`.

Previously, only udp- and mpi-conduits had `gasnetc_atttach()` logic
which departed from the template.  In both cases, the difference was
the presence of a `AMLOCK/UNLOCK` pair bracketing the call to
`gasneti_amregister_legacy()`.  This has been removed, since it did
not actually provide any necessary serialization.  The serialization
with *is* desired is relative to calls to `gex_EP_RegisterHandlers()`
and has been provided since PR#424.