mpi: remove locking around handler registration
This commit removes an `AMLOCK/AMUNLOCK` pair bracketing the call to
`gasneti_amregister_legacy()` in mpi-conduit's `gasnetc_attach()`, and
the error handling logic is adjusted accordingly.

There was already no analogous locking in `gex_EP_RegisterHandlers()`,
and no reason to believe (based on the implementation of
`AM_SetHandler()`) that serialization is necessary beyond that imposed
on the client.