ibv: Initial native implementation of NPAM Long
This commit provides an initial (complete/correct) native ibv-conduit
implementation of `gex_AM_{Prepare,Commit}{Request,Reply}Long()` and
corresponding adjustments to the max-payload queries.

The majority of the changes are generalization of the internal
functions `gasnetc_{prepare,commit}_medium()` to become
`gasnetc_{prepare,commit}_common()` with the addition of a `category`
argument.  This keeps a single version of the "main" code, and
makes the reasonable assumption that the optimizer will specialize the
inlined code using the manifest constant `category` in the callers.

The new callers of the common code are the "near public"
`gasnetc_AM_...()` entry points called in `gasnet_ammacros.h` after
insertion of `GASNETI_THREAD_GET` and handling of argument counts in
the `Commit` calls.  These are simple clones of the Medium versions
with the necessary mechanical changes for Medium-vs-Long.

Benchmarking of the client-allocated buffer case shows a small
degradation relative to the equivalent FPAM call.  This is expected
since there is no benefit to be gained for the transfer itself, and
the Prepare/Commit separation lengthens the critical path.

For the GASNet-allocated case, this implementation unconditionally
uses the "packed long" representation both in memory and for
communication, even for payload sizes where packed long is known to
under-perform in FPAM.  Testing on Summit, Wombat, Gomez and Dirac all
agree that this approach out-performs use of RDMA Put for the payload
(the implementation of which is not in this commit).  This result is
true for both ping-pong and flood tests, and for the full range of
payload sizes where a difference is observable (though small xfers are
too noisy).  Having already assembled the payload in registered memory
adjacent to the header, the "cost" of this approach is the `memcpy()`
on the target, and the "savings" is the injection of a single network
operation instead of two.  What this implies is that the cost of
injecting a second network operation (think "gap" in logP) exceeds the
cost of a 64KB `memcpy()` on the target on all four systems.