ibv: improve docs for firehose parameters
This commit rewrites the documentation for the `GASNET_FIREHOSE_*`
family of environment variables.

The intent of this rewrite is to better convey the connections between
the pairs of main/victim variables and between the sizing of the
memory pool and the registrations pool.

Additionally, an out-of-date reference to poor RDMA performance above
128k has been removed.

There are no intended changes to the documented semantics, just
clarification.