firehose: warn of "M" truncation
This commit adds logic to warn if the effective values of the firehose
"M" parameters are reduced due to insufficient region count to cover
their (default or explicit) values.

This does not change the default behavior of truncation to 4GB (3GB
main + 1GB victim), but rather just adds a warning when this occurs.
Avoiding the warning is the subject of another (near future) commit.