ibv: increase max firehose registrations
This commit adds logic to dynamically determine the maximum number of
pinned regions to be managed by firehose, rather than imposing a 32K
(compiled-in) limit.

The new logic imposes a 16M limit per host (enough for 2TB of physical
memory with defaults).  This is the `max_mr` capability seen on
current mlx_5-generation HCAs surveyed by the author (Summit, Wombat,
Dirac/CX5, Gomez and Thing), and so is unlikely to represent the
actual limit imposed.

Subject to the 16M cap, the value used is based on HCA capabilities.
The full capabilities of the HCA are queried and then ibv-conduit
claims the fraction (GASNET_PHYSMEM_MAX / physical-memory).  This is a
heuristic motivated by the observation that GASNET_PHYSMEM_MAX is
already provided to reserve/partition a fraction of resources to other
runtimes (such as MPI) or for the system (as for a shared filesystem).
The value MIN(16M, scaled-HCA-capability) is the per-host value, and
each co-located process is given an equal share.

It remains possible to set firehose's "R" parameters to override the
value chosen by the new logic, including the 16M limit.