firehose: revise fh_getenv() for case of `0` value
This commit revises the calling convention for `fh_getenv()` to
allow for the possibility of any explicit zero value.  Prior to this
change, an unset or empty environment variable would result in a
zero return, rendering a missing setting and an user-provided zero
indistinguishable.  Now there is a third argument, `*is_dflt` which
is used to return non-zero iff the env var was unset or empty.