firehose: add auto-sizing of region size
This commit adds the ability to request automatic sizing of the
firehose region size by setting `GASNET_FIREHOSE_MAXREGION_SIZE=0`.

With this new setting, if the default 128KiB region size and the
available region count are insufficient to address the pinnable memory
on the system, then the region size will be increased (subject to an
upper bound in some cases), to avoid the alternative behavior in which
the addressable memory is reduced (with a warning).  This is NOT made
the new default (at least not yet) due to empirical evidence that an
increase in the region size can negatively impact application
performance.