NO MERGE AS IS - fix unreported bug
This commit fixes a bug identified in the course of PR#449
https://bitbucket.org/berkeleylab/gasnet/pull-requests/449/bug-4209-ibv-alc#comment-217559115

Once Bugzilla is back online, this issue needs a bug report,
and this commit should at least have this commit message revised
to reference it.

As reported in bug XXXX, the use of `gex_Event_WaitAll()` on an array
containing both the root event and alc leaf event from
`gex_RMA_PutNB()` has been seen to fail with the message:
```
   *** FATAL ERROR (proc 0): in gasnete_test_array() at [redacted]/gasnet_event.c:395: Invalid use of free eop
```

The problem was a logic error in the debug checking code in
`gasnete_event_check()` generating the message above for both "free"
and "pendingfree" states of the eop.  The pendingfree state was
created specifically for use by the logic in `gasnete_test_array()`
(backing `gex_Event_WaitAll()`) to place root events "in limbo" until
their corresponding leaf events (if any) can be processed as well.
However, as written the `gasnete_event_check()` call on the leaf would
observe the pendingfree state of its root event as erroneous.

The fix in this commit is to separate the treatment of "free" and
pendingfree (for both eops and iops) to allow for a leaf event to be
pendingree, while continuing to treat pendingfree root events as
erroneous.