Tune gasneti_pshm_jobrank_in_supernode()
This commit deploys an improved algorithm within the function
`gasneti_pshm_jobrank_in_supernode()` to ensure that the most
common/important cases are decided using a single comparison.

For the common case of a contiguous nbrhd, jobranks within the
caller's supernode are correctly identified by the first comparison,
versus two required previously.  The failure case still takes two
comparisons.

For the case of a discontinuous nbrhd, one comparison is still
sufficient to identify "self" with the remaining cases requiring
a total of three comparisons.