ibv: revise protocol for piggy-backed credits
This commit revises the wire protocol for handling of piggy-backed
credits to use 8 of the 16 bits of immediate data space freed by the
allocation of a full 32-bit header field for the source rank.  This
replaces use of an "extra" argument which, when present encoded both
the credits and the true argument count.

This change significantly improves the readability of the AM header
sizing/generation logic, especially with respect to NPAM.  The crucial
paths for AM header processing (both on send and receive) should be
shorter as a result.