SQUASH/FIXUP - Initial support for HIP as a Memory Kind.
NOTE FOR REVIEW:

This commit reorders the content of gasnet_hip.c to be more logical,
as requested in PF review.

One function prototype is added, and some comments added or modified.
Other than those, this commit is simply code movement.
There are NO changes to function bodies, types, etc.