ibv: factor and improve "weak counter" abstraction
There are multiple places in ibv-conduit where we wish to round-robin
over multiple HCAs or QPs.  These have been written with a "benign"
thread race in which read-modify-write is non-atomic.  This is
expected to be more scalable than an atomic counter with little or no
negative impact from having multiple threads selecting the same value
concurrently due to the race.

This commit factors this abstraction and improves it by adding cache
padding around the counter in multi-threaded builds (including SEQ
builds with the AM receive thread enabled).  This padding eliminates
false sharing and yields measurable improvements in `testcontend`.