barrier: cache padding in rdma-dissem barriers
This commit reorganizes fields in the struct used by the reference,
aries- and ibv-conduit implementations of an RDMA dissemination
barrier.  The read/write field are separated from the read-only ones
and appropriate cache padding is added.

Of particular importance is the addition of padding between the lock
and the data it protects.  Without this, failing "try" traffic on the
lock was prone to repeated thrash the data.

Additionally, unnecessary duplication in aries-conduit has been
eliminated.