ibv: poll serialization by the AM receive thread
This commit implements a new `GASNET_RCV_THREAD_POLL_MODE` environment
variable, documented in the ibv-conduit `README`.  This setting
selects the polling mode of the AM receive thread.

The connection progress thread has not been modified to opt-in to
serialization, because when that thread is active all
connection-related completions occur on a distinct Cq that is *never*
polled by any other thread.