ibv: enable poll serialization in progress threads
This commit revises the implementation of ibv-conduit's progress
threads to allow for two forms of participation in the polling
serialization observed by client threads, in addition to the previous
non-participation.  One is participation as an equal peer with
the client threads, while the other is an exclusive mode in which the
progress thread acquires the serialization semaphore when started and
releases it when terminated.

This provides new capabilities for a given progress thread, but does
not enable it.  That will occur in a subsequent commit.