Simplify implementation of GEX_FLAG_AD_{ACQ,REL}
This commit resolves bug 4264:
   simplify implementation of GEX_FLAG_AD_{ACQ,REL}

The "meat" of this commit is the change in gasnet_ratomic.h from an
expression involving two `?:` operations and an `|`, by a single `&`.
The remainder of the commit supports that change by ensuring equallity
between the respective `ACQ` and `REL` constants in the `GEX_FLAG_AD_`
and `GASNETI_ATOMIC_` families.

The textually largest change is a change to several `GEX_FLAG_*`
constants.  To enable the desired alignment between constants, the
following changes are made:

+ `GEX_FLAG_AD_{ACQ,REL}` have been moved to the "1" and "2" positions
+ `GEX_FLAG_{SELF,PEER}_SEG` have shifted up by two positions to avoid
  conflicts with the new values of `GEX_FLAG_AD_{ACQ,REL}`.
+ `GEX_FLAG_AD_MY_{RANK,NBRHD}` and `GEX_FLAG_ENABLE_LEAF_LC` have
  also shifted up two positions, due to the new positions for the
  segment assertion flags.

Additionally, to avoid ever breaking the newly established equalities,
comments have been added to both sets of definitions and a pair if
static assertions have been added.