coll: fix a minor race in tree_geom_fetch
This commit relocates a previously unlocked read of a linked list head
to take place within the mutex-protected critical section.  Prior to
this change, a race in which two (or more) callers observed an empty
geom cache on entry would result in each caller initializing the cache
rather than appending to it.  The result would have been a silent
memory leak of the entires created by these threads other than the
last one to obtain the lock.  However, there was no correctness issue.

Additionally, availability of the list head has been used to make a
minor simplification to the list initialization code.