coll/reduce: generalize "eager" reductions
This commit generalizes the `reduce_BinomialEager` and
`reduce_BinomialEagerSeg` algorithms for ReduceToOne to optionally
perform a broadcast (over the same binomial tree as the reduction).
This is done if and only if `root == GEX_RANK_INVALID`, which will
soon be used internally as a means to request that these ReduceToOne
algorithms perform a ReduceToAll.