testreduce: add -p option for pollers
This commit extends `testreduce` with `-p N` option, similar to that
in `testam` and `testcoll`.  In a PAR build this spawns one or more
polling threads to run concurrently with the main thread.

This change is capable of reproducing bug 4266 on all conduits, where
previously it could only be reproduced using ibv-conduit's AM receive
thread.