Implement new GASNETI_NBRHD_{MAPPED,AUXSEG}*()
This commit implements the following six macros:
  `GASNETI_NBRHD_{MAPPED,AUXSEG}{,_ADDR,_ADDR_OR_NULL}()`

These are to replace `GASNETI_NBRHD_LOCAL_ADDR{,_OR_NULL}()`
(which are now deprecated) and supplement `GASNETI_NBRHD_LOCAL()`
(which remains available and _not_ deprecated).

The "MAPPED" family operate on a `(tm,rank)` which specify a remote
endpoint with a bound segment which is load/store accessible (e.g.
cross-mapped host memory), where the LOCAL family it is replacing was
not sensitive to the remote endpoint.  These do not accept addresses
in the aux segment, where the LOCAL macros accept(ed) both client and
aux segment addresses (at the cost of a comparison and branch).

The "AUXSEG" family fills the gap left by the inability of "MAPPED" to
process aux segment addresses.

The "LOCAL_ADDR" and "LOCAL_ADDR_OR_NULL" had *already* lost their
ability to process aux segment addresses earlier in the course of the
pull request containing this work, after determining that no aux seg
operations using them exist in the current code.  Therefore they are
retained (for the moment) as aliases for the "MAPPED" macros.
Subsequent commits will perform the renames and remove the aliases.