ucx: revise segment exchange logic
This commit reimplements `gasnetc_segment_exchange()` from scratch,
but retains a reduced version of its predecessor as
`gasnetc_segment_exchange_aux()`.

Previously the logic for exchange of registration keys for the aux and
at most one client segment were both handled in a combined
`gasnetc_segment_exchange()`.  Though that combined functionality was
written (see 6917dd1a) with the expectation of later extending to
multi-EP, that turned out to be impractical.  So, those two operations
have been separated, with several improvements enabled by the split.

The main improvements for the client segment exchange are as follows.
First, the signature is now the same as other conduits, making it
usable as a complete multi-ep-capable conduit-specific hook for use in
`gex_EP_PublishBoundSegment()`.  Second, use of the internal function
`gasneti_blockingRotatedExchangeV()` has eliminated the need to pad
elements to equal length, in turn removing the need for a MAX()
reduction over the variable rkey length.  Finally, this exchange now
communicates the "Location" of the EP rather than the address and
length of the segment (which are available in the segment data
structures).  This replaces 16 bytes (per segment) on the wire with
just 8 bytes, and when Publish is revised to merge conduit-independent
and conduit-dependent comms, these 8 bytes will be eliminated as well.
Additional savings come from use of just 2 bytes for the rkey length
(down from 8, and 1 might be sufficient).

The exchange of the aux segment rkeys is similarly simplified to
eliminate 8 bytes per process on the wire by finding the address and
length of the aux segment in existing data structures.  Removing the
generality added previously to to handle both aux and client segments
has significantly simplified the logic.  Unfortunately, there is no
alternative, available for use at bootstrap time, to the reduction and
padding to max length of an rkey.