Upgrade gasneti_hosthash()
This commit replaces the home-grown digest algorithm in
`gasneti_hosthash()` with the FVN-1a hash function.  This
non-cryptographic hash provides for significantly better
collision resistance than the code it replaces.

This change is work toward resolving bug 4297, by ensuring
that the `gasneti_hosthash()` result is a viable replacement
for (a properly implemented) `gethostid()`.