Use Cray PMI to validate nodemap
The Cray implementation of PMI includes an extension to return the
number of processes running on the caller's node/host.  This can be
used as a check on the value discovered by `gasneti_nodemapInit()`.
This is validation is done in (only) the first call to
`bootstrapSNodeBroadcast()` which, when PSHM is enable, will occur in
`gasneti_pshm_init()`, soon after nodemap construction.

This can improve the resolution of bug 4297 by providing a very clear
error when `gethostid()` has reported multiple distinct values for a
single host on a system running Cray PMI.  This replaces an obscure
failure of `gasneti_assert_always(i != count)` in the same first call
to `bootstrapSNodeBroadcast()` under the same conditions.