Use Cray PMI to validate nodemap
The Cray implementation of PMI includes an extension to return the
number of processes running on the caller's node/host.  This can be
used as a check on the value discovered by `gasneti_nodemapInit()`.
This is validation is done in (only) the first call to
`bootstrapSNodeBroadcast()` which, when PSHM is enable, will occur in
`gasneti_pshm_init()`, soon after nodemap construction.

This improves the resolution of bug 4297.