ucx: resolve bugs 4165 and 4280
This commit removes the `gasnetc_req_wait()` call (a stall for local
completion) from the end of AM injection, to resolve two related bugs.

Bug 4165 "ucx: AMMediumRequest bandwidth problems" is resolved by the
lack of a stall at the end of AM Request injection.  Since this was a
stall for remote completion when payloads were not delivered by an
eager protocol, it was crippling to Medium flood bandwidth.

Bug 4280 "ucx: investigate use of is_sync in AM injection" began as a
request to investigate the necessity for the `gasnetc_req_wait()`,
because initial attempts to remove it without understanding its
purpose had been observed to lead to hangs in testing.  As described
in comment #3 in that bug report, it turns out that the stall was
preventing deadlock in `gasnetc_am_req_get()` -- but that was better
addressed by prior updates to that function to poll properly.  So
there is no longer any need for `is_sync`.